+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonerfinancepocapi/models/sync/SyncOffenderTransactionRequest.kt
@@ -13,7 +13,7 @@ data class SyncOffenderTransactionRequest(
     example = "19228028",
     required = true,
   )
-  val transactionId: Int,
+  val transactionId: Long,
 
   @Schema(
     description = "A unique identifier for this synchronization request. This can be used for idempotency or tracing.",
